fix compare warning

+++ test/RandomTest.cpp
@@ -53,7 +53,7 @@
     RandomString rs;                                                           \
     for (int i = 0; i < TEST_MAX; i++) {                                       \
       std::string s = rs.f(i + 1);                                             \
-      for (int j = 0; j < s.length(); j++) {                                   \
+      for (int j = 0; j < (int)s.length(); j++) {                              \
         REQUIRE(g(s[j]));                                                      \
       }                                                                        \
     }                                                                          \
